NXDRIVE-2201: Move the NXDRIVE-1033 test in its own test file
