+++ docs/changes/4.4.4.md
@@ -53,6 +53,7 @@ Release date: `2020-xx-xx`
 - [NXDRIVE-2189](https://jira.nuxeo.com/browse/NXDRIVE-2189): Move non-Python test files to specific folders
 - [NXDRIVE-2199](https://jira.nuxeo.com/browse/NXDRIVE-2199): Use the FileManager importer in tests
 - [NXDRIVE-2200](https://jira.nuxeo.com/browse/NXDRIVE-2200): Fix tests from `test_special_characters.py`
+- [NXDRIVE-2201](https://jira.nuxeo.com/browse/NXDRIVE-2201): Move the NXDRIVE-1033 test in its own test file
 
 ## Docs
